mtd::StubGLRenderingProvider: Pass through Buffer in as_texture if possible.
Some of the `Buffer`s used in the tests are real SHM buffers, and the tests
rely on their behaviour being authentic - notably, they need `::bind()` to
trigger the consumed event, and (for WLCS) they need to actually touch the
memory.

Have `StubGLRenderingProvider` just downcast the passed-in `Buffer` if possible,
and synthesise a new one if not.