gbm-kms: Fix KMS framebuffer lifecycle.
We need up to *three* framebuffers here:
1. The one that's onscreen
2. The one that we've submitted the hardware, waiting for it to flip onscreen
3. The one that we're about to submit to the hardware.

In the common case (not-clone-mode) we'll wait for (2) to become (1) before
returning from `post()`. Due to clones not necessarily being synchronised
to the same clock we don't want to wait for all of them to flip in `post()`,
so we potentially need all three FBs.

We should revisit to check that the reasoning still applies when switching to Atomic KMS.